+++ museek/config/process_uhf_band.py
@@ -1,7 +1,12 @@
+"""MuSEEK UHF data processing pipeline configuration template."""
+
 from ivory.utils.config_section import ConfigSection
 
+# --- Pipeline Configuration ---
 Pipeline = ConfigSection(
+    # Order of the plugins to run. Comment out to skip steps.
     plugins=[
+        # --- Pre self-calibration steps ---
         "museek.plugin.in_plugin",
         "museek.plugin.noise_diode_flagger_plugin",
         "museek.plugin.known_rfi_plugin",
@@ -14,66 +19,86 @@ Pipeline = ConfigSection(
         "museek.plugin.noise_diode_plugin",
         "museek.plugin.gain_calibration_plugin",
         "museek.plugin.aoflagger_postcalibration_plugin",
+        "museek.plugin.inpainting_mapmaking_plugin",
+        # --- Self-calibration steps (disable by default) ---
+        # Disable by default as running these steps through currently requires combining
+        # maps from multiple blocks using map_making_inpainting_combineblocks.ipynb
+        # before inpainting_mapmaking and gain_selfcalibration steps
         #'museek.plugin.gain_selfcalibration_plugin',
         #'museek.plugin.aoflagger_postselfcalibration_plugin',
+        #'museek.plugin.inpainting_mapmaking_selfcali_plugin',
         #'museek.plugin.aoflagger_cross_plugin',
         #'museek.plugin.single_dish_calibrator_plugin',
         #'museek.plugin.zebra_remover_plugin',
         #'museek.plugin.apply_external_gain_solution_plugin',
     ],
-    # context=os.path.join('/idia/users/wkhu/calibration_results/noise_diode_2d/', '1675632179/gain_calibration_plugin.pickle')
+    # --- Resume behaviour ---
+    # Modify and uncomment `context` parameter to load a saved context pickle file
+    # required to continue from desired step.
+    # context="/path/to/context/folder/plugin_name.pickle"
 )
 
+# Load and extract visibilities, flags, and weights from correlator data
 InPlugin = ConfigSection(
-    block_name="1675632179",  # observation time stamp
-    # receiver_list=['m000h','m000v','m012h','m012v','m024h','m024v','m036h','m036v'],
-    receiver_list=None,  # receivers to be processed, `None` means all available receivers is used
-    token=None,  # archive token
-    data_folder="/idia/projects/meerklass/MEERKLASS-1/SCI-20220822-MS-01/",  # only relevant if `token` is `None`
-    # data_folder='/idia/projects/hi_im/SCI-20230907-MS-01/',  # only relevant if `token` is `None`
-    force_load_auto_from_correlator_data=True,  # if `True`, the local `cache` folder is ignored
-    force_load_cross_from_correlator_data=True,  # if `True`, the local `cache` folder is ignored
-    do_save_visibility_to_disc=False,  # if `True`, the extracted visibilities, flags and weights are stored to disc for quicker access
-    do_store_context=False,
-    context_folder=None,  # directory to store results, if `None`, 'results/' is chosen
+    # --- Parameters to change ---
+    block_name="1675632179",  # Block number of the observation
+    data_folder="/idia/projects/meerklass/MEERKLASS-1/SCI-20220822-MS-01/",  # Directory containing MVF4 correlator data
+    # --- Optional and fixed parameters ---
+    receiver_list=None,  # Set to `None` to use all receivers or provide a list, e.g.
+    # receiver_list=['m000h','m000v','m012h','m012v',...],
+    token=None,  # Token for loading data directly from SARAO archive
+    force_load_auto_from_correlator_data=True,  # If `True`, ignore local cache folder and read from correlator data
+    force_load_cross_from_correlator_data=True,  # If `True`, ignore local cache folder and read from correlator data
+    do_save_visibility_to_disc=False,  # If `True`, extracted save visibilities, flags and weights to local cache folder
+    do_store_context=False,  # If `True`, Save the context pickle file from this step
+    context_folder=None,  # Directory to store the context files "./results" is used if `None`
 )
 
-
+# Flag known RFI bands
 KnownRfiPlugin = ConfigSection(
+    # --- Parameters to change ---
     gsm_900_uplink=None,
     gsm_900_downlink=(925, 960),
     gsm_1800_uplink=None,
-    # gps=(1170, 1390),
-    # extra_rfi=[(1524, 1630)],
     gps=None,
-    extra_rfi=[(765, 775), (801, 811), (811, 821)],  # Vodacom  # MTN  # Telkom
+    extra_rfi=[(768, 778), (801, 811), (811, 821)],  # Vodacom, MTN and Telkom bands
 )
 
-RawdataFlaggerPlugin = ConfigSection(flag_lower_threshold=5.0, do_store_context=False)
+# Flag raw data with values below a minimum threshold
+RawdataFlaggerPlugin = ConfigSection(
+    # --- Parameters to change ---
+    flag_lower_threshold=5.0,  # Minimum threshold in raw correlator unit
+    do_store_context=False,
+)
 
-
+# Split calibrator tracks and scan data
 ScanTrackSplitPlugin = ConfigSection(
-    do_delete_unsplit_data=True, do_store_context=False
+    # --- Parameters to change ---
+    do_delete_unsplit_data=True,  # Delete original data after splitting
+    do_store_context=False,
 )
 
-
+# Flag point source using a catalog
 PointSourceFlaggerPlugin = ConfigSection(
+    # --- Parameters to change ---
     n_jobs=26,
     verbose=0,
     point_source_file_path="/idia/projects/meerklass/MEERKLASS-1/uhf_data/OT2023/radio_source_catalog/",
-    beam_threshold=1.0,  # times of the beam size around the point source to be masked
-    point_sources_match_flux=5.0,  # flux threshold above which the point sources are selected, [Jy]
-    point_sources_match_raregion=30.0,  # the ra distance to the median of observed ra to select the point sources, [deg]
-    point_sources_match_decregion=10.0,  # the dec region to the median of observed dec to select the point sources [deg]
-    beamsize=57.5,  # the beam fwhm used to smooth the Synch model [arcmin]
-    beam_frequency=1500.0,  # reference frequency at which the beam fwhm are defined [MHz]
+    beam_threshold=1.0,  # Times of the beam size around the point source to be masked
+    point_sources_match_flux=5.0,  # Flux threshold above which the point sources are selected, [Jy]
+    point_sources_match_raregion=30.0,  # The ra distance to the median of observed ra to select the point sources, [deg]
+    point_sources_match_decregion=10.0,  # The dec region to the median of observed dec to select the point sources [deg]
+    beamsize=57.5,  # The beam fwhm used to smooth the Synch model [arcmin]
+    beam_frequency=1500.0,  # Reference frequency at which the beam fwhm are defined [MHz]
 )
 
-
+# First round of flagging using AOflagger algorithm
 AoflaggerPlugin = ConfigSection(
+    # --- Parameters to change ---
     n_jobs=26,
     verbose=0,
-    mask_type="vis",  # the data to which the flagger will be applied, ['vis', 'flag_fraction', 'rms', 'inverse', 'inverse_timemedian']
+    mask_type="vis",  # The data to which the flagger will be applied, choices are:
+    # ['vis', 'flag_fraction', 'rms', 'inverse', 'inverse_timemedian']
     first_threshold=0.1,  # First threshold value
     threshold_scales=[0.5, 0.55, 0.62, 0.75, 1],
     smoothing_kernel=(
@@ -84,7 +109,7 @@ AoflaggerPlugin = ConfigSection(
     struct_size=(
         3,
         3,
-    ),  # size of struct for dilation in time and frequency direction [pixels]
+    ),  # Size of struct for dilation in time and frequency direction [pixels]
     channel_flag_threshold=0.6,
     time_dump_flag_threshold=0.6,
     flag_combination_threshold=1,
@@ -92,6 +117,7 @@ AoflaggerPlugin = ConfigSection(
 )
 
 AoflaggerCrossPlugin = ConfigSection(
+    # --- Parameters to change ---
     n_jobs=26,
     verbose=0,
     mask_type="vis",  # the data to which the flagger will be applied, ['vis', 'flag_fraction', 'rms', 'inverse', 'inverse_timemedian']
@@ -113,6 +139,7 @@ AoflaggerCrossPlugin = ConfigSection(
 )
 
 AoflaggerSecondRunPlugin = ConfigSection(
+    # --- Parameters to change ---
     n_jobs=13,
     verbose=0,
     mask_type="vis",  # the data to which the flagger will be applied, ['vis', 'inverse'] for 1d aoflagger
@@ -128,10 +155,12 @@ AoflaggerSecondRunPlugin = ConfigSection(
     time_dump_flag_threshold=0.6,
     flag_combination_threshold=1,
     do_store_context=False,
+    context_folder=None,  # New path to save output, if `None`, using existing path in the input context file
 )
 
 
 AntennaFlaggerPlugin = ConfigSection(
+    # --- Parameters to change ---
     elevation_std_threshold=1e-2,  # standard deviation threshold of individual dishes elevation in degrees
     elevation_threshold=0.1,  # time points with elevation reading deviations exceeding this threshold are flagged
     outlier_threshold=0.1,  # antenna outlier threshold [degrees]
@@ -141,6 +170,7 @@ AntennaFlaggerPlugin = ConfigSection(
 
 
 NoiseDiodePlugin = ConfigSection(
+    # --- Parameters to change ---
     n_jobs=13,
     verbose=0,
     flag_combination_threshold=1,
@@ -154,6 +184,7 @@ NoiseDiodePlugin = ConfigSection(
 
 
 GainCalibrationPlugin = ConfigSection(
+    # --- Parameters to change ---
     cali_method="corr",  # method to do the calibration 'corr' or 'rms'
     synch_model=[
         "s1"
@@ -172,10 +203,11 @@ GainCalibrationPlugin = ConfigSection(
     nd_window_movingmedian=20,  # The size of the window for the moving median calculation for frequency spectrum of noise diode signal
     nd_gausm_sigma=20,  # The size of the window for the Gaussian Smooth of Noise Diode Excess frequency spectrum
     do_delete_auto_data=False,  # switch that determines wether the raw auto data should be deleted after calibration
+    context_folder=None,  # New path to save output, if `None`, using existing path in the input context file
 )
 
-
 AoflaggerPostCalibrationPlugin = ConfigSection(
+    # --- Parameters to change ---
     n_jobs=13,
     verbose=0,
     first_threshold_rms=0.1,  # First threshold value
@@ -204,11 +236,38 @@ AoflaggerPostCalibrationPlugin = ConfigSection(
     zscore_antenatempflag_threshold=5.0,  # threshold for flagging the antennas based on their average temperature using modified zscore method
     do_store_context=True,
     do_delete_auto_data=False,  # switch that determines wether the raw auto data, flags and weights should be deleted after calibration
-    new_output_path=None,  # new path to save output, if `None`, using old output_path
+    ### change the new_output_path if the output_path that was saved in the loaded plugin (pickle) is not existing or not what you want
+    context_folder=None,  # New path to save output, if `None`, using existing path in the input context file
+    gsm_900_uplink=None,
+    gsm_900_downlink=(925, 960),
+    gsm_1800_uplink=None,
+    gps=None,
+    extra_rfi=[(768, 778), (801, 811), (811, 821)],
 )
 
 
+InpaintingMapmakingPlugin = ConfigSection(
+    ### Parameters that are required to be changed
+    x_crval=165.0,  # map center x(ra) [deg], this value is for DESI 2
+    y_crval=-1.5,  # map center y(dec) [deg], this value is for DESI 2
+    # --- Parameters to change ---
+    n_jobs=13,
+    verbose=0,
+    threshold_MHz=30.0,  # if a long continuous frequency region is masked, this timestamp will be totally masked [MHz]
+    inpainting_window=20.0,  # inpainting the masked regions by fitting a polynomial using +-inpainting_window of the unmasked data around the masked regions [MHz]
+    inpainting_polydeg=6,  # the degree of polynomials fit in inpainting
+    mask_antnum_threshold=10,  # masking the pixels where <=mask_antnum_threshold antennas contributes
+    pix_reso=0.5,  # map resolution [deg]
+    x_range=35,  # map range (+-x_range) in x(ra) [deg]
+    y_range=10,  # map range (+-y_range) in y(dec) [deg]
+    do_store_context=True,
+)
+
 GainSelfCalibrationPlugin = ConfigSection(
+    ### Parameters that are required to be changed
+    map_path="/idia/projects/meerklass/MEERKLASS-2/UHF/DESI_2/",  # directory of the input map -- update this to match your own directory.
+    map_name="map_making_DESI_2_model.pkl",  # name of the input map -- update this to match your own filename
+    # --- Parameters to change ---
     n_jobs=13,
     verbose=0,
     frequency_high=1015.0,  # high frequency cut for the scan data [MHz]
@@ -222,8 +281,6 @@ GainSelfCalibrationPlugin = ConfigSection(
     nd_window_movingmedian=20,  # The size of the window for the moving median calculation for frequency spectrum of noise diode signal
     nd_gausm_sigma=20,  # The size of the window for the Gaussian Smooth of Noise Diode Excess frequency spectrum
     do_delete_auto_data=True,  # switch that determines wether the raw auto data should be deleted after calibration
-    map_path="/idia/projects/meerklass/MEERKLASS-2/UHF/DESI_2/",  # directory of the input map
-    map_name="map_making_DESI_2_model.pkl",  # name of the input map
     baseline_polyfit_deg=2,  # degree of the polynomials used for fitting time baseline in raw data
     gain_polyfit_deg=2,  # degree of the polynomials used for fitting gain
     model_polyfit_deg=6,  # degree of the polynomials used for fitting model
@@ -233,6 +290,7 @@ GainSelfCalibrationPlugin = ConfigSection(
 
 
 AoflaggerPostSelfCalibrationPlugin = ConfigSection(
+    # --- Parameters to change ---
     n_jobs=13,
     verbose=0,
     first_threshold_rms=0.1,  # First threshold value
@@ -261,14 +319,32 @@ AoflaggerPostSelfCalibrationPlugin = ConfigSection(
     zscore_antenatempflag_threshold=5.0,  # threshold for flagging the antennas based on their average temperature using modified zscore method
     do_store_context=True,
     do_delete_auto_data=True,  # switch that determines wether the raw auto data should be deleted after calibration
+    gsm_900_uplink=None,
+    gsm_900_downlink=(925, 960),
+    gsm_1800_uplink=None,
+    gps=None,
+    extra_rfi=[(768, 778), (801, 811), (811, 821)],
 )
 
-
-ApplyExternalGainSolutionPlugin = ConfigSection(
-    gain_file_path="/home/amadeus/Documents/fix/postdoc_UWC/work/MeerKLASS/calibration/download/level2/"
+InpaintingMapmakingSelfcaliPlugin = ConfigSection(
+    # --- Parameters to change ---
+    n_jobs=13,
+    verbose=0,
+    threshold_MHz=30.0,  # if a long continuous frequency region is masked, this timestamp will be totally masked [MHz]
+    inpainting_window=20.0,  # inpainting the masked regions by fitting a polynomial using +-inpainting_window of the unmasked data around the masked regions [MHz]
+    inpainting_polydeg=6,  # the degree of polynomials fit in inpainting
+    mask_antnum_threshold=10,  # masking the pixels where <=mask_antnum_threshold antennas contributes
+    do_store_context=True,
 )
 
-ZebraRemoverPlugin = ConfigSection(
-    reference_channel=3000,
-    zebra_channels=range(350, 498),
-)
+# --- Current not used, incomplete or obsolete plugins
+# ApplyExternalGainSolutionPlugin = ConfigSection(
+#     # --- Parameters to change ---
+#     gain_file_path="/home/amadeus/Documents/fix/postdoc_UWC/work/MeerKLASS/calibration/download/level2/"
+# )
+
+# ZebraRemoverPlugin = ConfigSection(
+#     # --- Parameters to change ---
+#     reference_channel=3000,
+#     zebra_channels=range(350, 498),
+# )